Research-Report-Generator/requirements.txt
crewai
crewai-tools
python-dotenv
langchain-google-genai
duckduckgo-search
python-docx
google-generativeai